requirements.txt
opencv-python
numpy
shapely
tqdm
matplotlib
tensorflow
scikit-learn
seaborn
fastapi
uvicorn
python-multipart
pillow